requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
azure-common==1.1.27
azure-functions==1.7.0
azure-functions-durable==1.0.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==4.4.2
furl==2.1.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
multidict==5.1.0
numpy==1.20.3
orderedmultidict==1.0.1
pandas==1.2.4
Pillow==8.2.0
proglog==0.1.9
pycparser==2.20
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
